Update Homework/Homework/Program.cs
Понимаю твоё негодование, но, дал бы мне компилятор упростить (имею ввиду без обязательного указания что i - это целое число, так бы и сделал. Понятно что это мусор и так писать не нужно. Хз какой то конфликт с VS умолял прямую инициализацию ему указать

+++ Homework/Homework/Program.cs
@@ -189,10 +189,10 @@ class Program
         int indexOfNumb = 0;
         for (double i = 0; array.Length > i; i++)
         {
-            if (maxNumber < array[(int)i])
+            if (maxNumber < array[i])
             {
-                maxNumber = array[(int)i];
-                indexOfNumb = (int)i;
+                maxNumber = array[i];
+                indexOfNumb = i;
             }
         }
         Console.WriteLine(maxNumber);
